Refactor: Extract static file provider classes into dedicated files and reorganize `StaticWebAssetsRuntimeFileProvider` internals
- Moved `StaticWebAsset`, `StaticWebAssetPattern`, `NodeTraversalState`, `ManifestCandidate`, and related classes to separate files under `FileProviders.Static`.
- Updated namespace references in affected files and tests for consistency.
- Improved maintainability of `StaticWebAssetsRuntimeFileProvider` by restructuring internal implementations.
