+++ src/InfiniFrame.BlazorWebView/FileProviders/Static/ManifestCandidate.cs
@@ -1,0 +1,11 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+namespace InfiniFrame.BlazorWebView.FileProviders.Static;
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+internal sealed record ManifestCandidate(
+    string ManifestPath,
+    int BaseScore
+);
