+++ src/InfiniFrame.BlazorWebView/FileProviders/Static/ManifestDirectoryContents.cs
@@ -1,0 +1,16 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+using Microsoft.Extensions.FileProviders;
+
+namespace InfiniFrame.BlazorWebView.FileProviders.Static;
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+internal sealed class ManifestDirectoryContents(IReadOnlyList<IFileInfo> entries) : IDirectoryContents {
+    public bool Exists => true;
+
+    public IEnumerator<IFileInfo> GetEnumerator() => entries.GetEnumerator();
+
+    System.Collections.IEnumerator System.Collections.IEnumerable.GetEnumerator() => GetEnumerator();
+}
