+++ src/InfiniFrame.BlazorWebView/FileProviders/Static/NodeTraversalState.cs
@@ -1,0 +1,12 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+namespace InfiniFrame.BlazorWebView.FileProviders.Static;
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+internal sealed record NodeTraversalState(
+    StaticWebAssetNode Node,
+    int ConsumedSegments,
+    string PathPrefix
+);
