+++ src/InfiniFrame.BlazorWebView/FileProviders/Static/ScoredManifestCandidate.cs
@@ -1,0 +1,11 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+namespace InfiniFrame.BlazorWebView.FileProviders.Static;
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+internal sealed record ScoredManifestCandidate(
+    StaticWebAssetManifest Manifest,
+    int Score
+);
