+++ src/InfiniFrame.BlazorWebView/FileProviders/Static/StaticWebAssetPattern.cs
@@ -1,0 +1,18 @@
+﻿// ---------------------------------------------------------------------------------------------------------------------
+// Imports
+// ---------------------------------------------------------------------------------------------------------------------
+using JetBrains.Annotations;
+using System.Text.Json.Serialization;
+
+namespace InfiniFrame.BlazorWebView.FileProviders.Static;
+// ---------------------------------------------------------------------------------------------------------------------
+// Code
+// ---------------------------------------------------------------------------------------------------------------------
+[UsedImplicitly]
+internal sealed class StaticWebAssetPattern {
+    [JsonPropertyName("ContentRootIndex")]
+    public int ContentRootIndex { get; set; }
+
+    [JsonPropertyName("Pattern")]
+    public string Pattern { get; set; } = string.Empty;
+}
